// packages/typespec-test/test/todo/generated/typespec-ts/src/models/models.ts
// Copyright (c) Microsoft Corporation.
// Licensed under the MIT license.

export interface TodoPage {
  /** The items in the page */
  items: TodoItem[];
  pagination: {
    pageSize: number;
    totalSize: number;
    prevLink?: string;
    nextLink?: string;
  };
}

export interface TodoItem {
  /** The item's unique id */
  id: number;
  /** The item's title */
  title: string;
  /** User that created the todo */
  readonly createdBy: number;
  /** User that the todo is assigned to */
  ownedBy: number;
  /** A longer description of the todo item in markdown format */
  description: string;
  /** The status of the todo item */
  status: "NotStarted" | "InProgress" | "Completed";
  /** When the todo item was created. */
  readonly createdAt: Date;
  /** When the todo item was last updated */
  readonly updatedAt: Date;
  /** When the todo item was makred as completed */
  readonly completedAt: Date;
  labels: TodoLabel[];
}

export interface TodoLabelRecord {
  name: string;
  color?: string;
}

export interface TodoUrlAttachment {
  /** A description of the URL */
  description: string;
  /** The url */
  url: string;
}

export interface TodoItemPatch {
  /** The item's title */
  title?: string;
  /** User that the todo is assigned to */
  ownedBy?: number;
  /** A longer description of the todo item in markdown format */
  description?: string;
  /** The status of the todo item */
  status?: "NotStarted" | "InProgress" | "Completed";
}

export interface TodoFileAttachment {
  /** The todo item this is attached to */
  todoItemId: number;
  /** The file name of the attachment */
  filename: string;
  /** The media type of the attachment */
  mediaType: string;
  /** The url where the attachment can be downloaded from */
  url: string;
}

export interface User {
  /** An autogenerated unique id for the user */
  readonly id: number;
  /** The user's username */
  username: string;
  /** The user's email address */
  email: string;
  /**
   * The user's password, provided when creating a user
   * but is otherwise not visible (and hashed by the backend)
   */
  password: string;
}

export interface UserCreatedResponse {
  /** An autogenerated unique id for the user */
  readonly id: number;
  /** The user's username */
  username: string;
  /** The user's email address */
  email: string;
  /**
   * The user's password, provided when creating a user
   * but is otherwise not visible (and hashed by the backend)
   */
  password: string;
  /** The token to use to construct the validate email address url */
  token: string;
}

/** Alias for TodoLabel */
export type TodoLabel = string | string[] | TodoLabelRecord | TodoLabelRecord[];
/** Alias for TodoAttachment */
export type TodoAttachment = TodoFileAttachment | TodoUrlAttachment;